pyproject.toml
[project]
name = "api-auditoria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "pymysql>=1.1.0",
]
